# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SILICOFELLER Quantum Studio API",
        "description": "AI-augmented quantum hardware EDA platform — V2. DesignGraph → Constraints → Placement → Routing → DRC → Qiskit Metal → Tapeout.",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/qclang/parse": {
            "post": {
                "tags": [
                    "qclang"
                ],
                "summary": "Parse Qclang",
                "description": "Parse QCLang source and return AST + validation errors.",
                "operationId": "parse_qclang_api_qclang_parse_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Parse Qclang Api Qclang Parse Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/qclang/compile": {
            "post": {
                "tags": [
                    "qclang"
                ],
                "summary": "Compile Qclang",
                "description": "Parse, validate, and compile QCLang source.\n\nAuto-detects dialect:\n  • .qcl (chip{} braces) → full QChipLang compiler → supports qiskit_metal | json_ir | spice\n  • .qc  (chip…end)       → simple compiler → qiskit_metal only\n\nReturns GenerateResponse-compatible result.",
                "operationId": "compile_qclang_api_qclang_compile_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompileRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Compile Qclang Api Qclang Compile Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/qclang/save": {
            "post": {
                "tags": [
                    "qclang"
                ],
                "summary": "Save Qclang",
                "description": "Parse + save a QCLang file to a project.",
                "operationId": "save_qclang_api_qclang_save_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Save Qclang Api Qclang Save Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompileRequest": {
                "properties": {
                    "source": {
                        "type": "string",
                        "title": "Source"
                    },
                    "target": {
                        "type": "string",
                        "title": "Target",
                        "default": "qiskit_metal"
                    },
                    "target_freq_ghz": {
                        "type": "number",
                        "title": "Target Freq Ghz",
                        "default": 5.0
                    },
                    "substrate": {
                        "type": "string",
                        "title": "Substrate",
                        "default": "silicon"
                    },
                    "metal": {
                        "type": "string",
                        "title": "Metal",
                        "default": "aluminum"
                    },
                    "chip_size_mm": {
                        "type": "number",
                        "title": "Chip Size Mm",
                        "default": 10.0
                    },
                    "project_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Project Id"
                    }
                },
                "type": "object",
                "required": [
                    "source"
                ],
                "title": "CompileRequest"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ParseRequest": {
                "properties": {
                    "source": {
                        "type": "string",
                        "title": "Source"
                    },
                    "project_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Project Id"
                    }
                },
                "type": "object",
                "required": [
                    "source"
                ],
                "title": "ParseRequest"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/auth/token"
                    }
                }
            }
        }
    }
}